:truck: rename: Tag 도메인 생성에 따른 TagResponse 클래스 도메인 패키지 내로 이동
